Move cafe to package and refactoring
